调整部署结构

+++ jsoncpp/deploy.bat
@@ -5,15 +5,25 @@ SETLOCAL EnableDelayedExpansion
 set INIT_DIR=%1
 set REPO_URL=%2
 set REPO_VER=%3
-set GENERATOR=%4
+set COMPILER=%4
+set ARCH=%5
 echo INIT_DIR: %INIT_DIR%
 echo REPO_URL: %REPO_URL%
 echo REPO_VER: %REPO_VER%
-echo GENERATOR: %GENERATOR%
-pause
+echo COMPILER: %COMPILER%
+echo ARCH: %ARCH%
 
-rem generator with underline
-set generator_ul=%GENERATOR: =_%
+rem compiler with underline
+set compiler_ul=%COMPILER: =_%
+echo compiler_ul: %compiler_ul%
+
+if defined ARCH (
+	set generator="%COMPILER:"=% %ARCH:"=%"
+) else (
+	set generator="%COMPILER:"=%"
+)
+set generator_ul=%generator: =_%
+echo generator: %generator%
 echo generator_ul: %generator_ul%
 
 for /f "tokens=4* delims=/" %%i in (%REPO_URL%) do (
@@ -29,13 +39,16 @@ set repo_path_win=%repo_path:/=\%
 echo repo_path: %repo_path%
 echo repo_path_win: %repo_path_win%
 
-set install_prefix=%repo_path:"=%/%generator_ul:"=%
+if defined ARCH (
+	set install_prefix=%repo_path:"=%/%compiler_ul:"=%/%ARCH:"=%
+) else (
+	set install_prefix=%repo_path:"=%/%compiler_ul:"=%/Win32
+)
 set install_prefix_win=%install_prefix:/=\%
 echo install_prefix: %install_prefix%
 echo install_prefix_win: %install_prefix_win%
-pause
 
-mkdir %repo_path_win%
+if not exist %repo_path_win% mkdir %repo_path_win%
 cd /d %repo_path_win%
 
 if not exist %repo_name% (
@@ -44,41 +57,34 @@ if not exist %repo_name% (
 	cd %repo_name%
 	git checkout -b %REPO_VER% %REPO_VER%
 	cd ..
-	cd
 )
-pause
-cd
-echo cd %repo_name%
 cd %repo_name%
-pause
-mkdir build
-pause
+if not exist build mkdir build
 cd build
-pause
 
-rd /q /s %generator_ul%
+if exist %generator_ul% rd /q /s %generator_ul%
 :wait_rd_generator_ul
-timeout 1
+timeout 2
 if exist %generator_ul% (
 	goto wait_rd_generator_ul
 )
-mkdir %generator_ul%
+if not exist %generator_ul% mkdir %generator_ul%
 cd %generator_ul%
 cmake -G %GENERATOR% ../.. -DJSONCPP_WITH_CMAKE_PACKAGE=ON -DCMAKE_INSTALL_PREFIX=%install_prefix%
-pause
-echo rd /q /s %install_prefix_win%
-rd /q /s %install_prefix_win%
-pause
+
+if exist %install_prefix_win% rd /q /s %install_prefix_win%
+
 cmake --build . --config Debug --target install || goto :error
-pause
-mkdir %install_prefix_win%\Debug
+if not exist %install_prefix_win%\Debug mkdir %install_prefix_win%\Debug
 move %install_prefix_win%\include %install_prefix_win%\Debug\include
 move %install_prefix_win%\lib %install_prefix_win%\Debug\lib
-pause
+timeout 2
+
 cmake --build . --config Release --target install || goto :error
-mkdir %install_prefix_win%\Release
+if not exist %install_prefix_win%\Release mkdir %install_prefix_win%\Release
 move %install_prefix_win%\include %install_prefix_win%\Release\include
 move %install_prefix_win%\lib %install_prefix_win%\Release\lib
+timeout 2
 
 :error
 pause
